Split agentHostSessionConfigPicker into base, mobile, and contribution files
Mirrors the AgentHostModePicker / MobileAgentHostModePicker structure:

* agentHostSessionConfigPicker.ts — base class + helper exports only
  (IConfigPickerItem, getConfigIcon). All Action2 registrations, the
  contribution, and the inline MobileAgentHostSessionConfigPicker moved
  out.
* mobile/mobileAgentHostSessionConfigPicker.ts (new) — phone subclass
  with the four mobile overrides (_orderProperties, _shouldRenderProperty,
  _isReadOnlyChip, _showPicker) and the bottom-sheet helper.
* agentHostSessionConfigPicker.contribution.ts (new) —
  AgentHostSessionConfigPickersContribution (renamed plural) plus the
  Action2 registrations and PickerActionViewItem. The session-config
  picker now picks the mobile subclass at view-item construction time via
  the same factory pattern used for the mode picker.

The isPhoneLayout guard added in #316151 to _orderProperties / _showPicker
is no longer needed and has been removed — the mobile subclass is only
instantiated on phone viewports via the factory, so its overrides are
unconditionally mobile.
