+++ src/vs/sessions/contrib/providers/agentHost/browser/agentHostSessionConfigPicker.ts
@@ -9,63 +9,28 @@ import { Gesture, EventType as TouchEventType } from '../../../../../base/browse
 import { renderIcon } from '../../../../../base/browser/ui/iconLabel/iconLabels.js';
 import { ActionListItemKind, IActionListDelegate, IActionListItem } from '../../../../../platform/actionWidget/browser/actionList.js';
 import { IActionWidgetService } from '../../../../../platform/actionWidget/browser/actionWidget.js';
-import { BaseActionViewItem } from '../../../../../base/browser/ui/actionbar/actionViewItems.js';
 import { Delayer } from '../../../../../base/common/async.js';
 import { Codicon } from '../../../../../base/common/codicons.js';
 import { MarkdownString } from '../../../../../base/common/htmlContent.js';
-import { Disposable, DisposableMap, DisposableStore, IDisposable } from '../../../../../base/common/lifecycle.js';
+import { Disposable, DisposableMap, DisposableStore } from '../../../../../base/common/lifecycle.js';
 import { autorun } from '../../../../../base/common/observable.js';
 import Severity from '../../../../../base/common/severity.js';
 import { ThemeIcon } from '../../../../../base/common/themables.js';
-import { localize, localize2 } from '../../../../../nls.js';
-import { IActionViewItemService } from '../../../../../platform/actions/browser/actionViewItemService.js';
-import { Action2, MenuId, registerAction2 } from '../../../../../platform/actions/common/actions.js';
+import { localize } from '../../../../../nls.js';
 import { IConfigurationService } from '../../../../../platform/configuration/common/configuration.js';
-import { ContextKeyExpr, IContextKeyService } from '../../../../../platform/contextkey/common/contextkey.js';
+import { IContextKeyService } from '../../../../../platform/contextkey/common/contextkey.js';
 import { IDialogService } from '../../../../../platform/dialogs/common/dialogs.js';
-import { IInstantiationService } from '../../../../../platform/instantiation/common/instantiation.js';
 import { ITelemetryService } from '../../../../../platform/telemetry/common/telemetry.js';
 import type { SessionConfigPropertySchema, SessionConfigValueItem } from '../../../../../platform/agentHost/common/state/protocol/commands.js';
 import { ChatConfiguration } from '../../../../../workbench/contrib/chat/common/constants.js';
-import { ChatContextKeyExprs } from '../../../../../workbench/contrib/chat/common/actions/chatContextKeys.js';
-import { IWorkbenchContribution, registerWorkbenchContribution2, WorkbenchPhase } from '../../../../../workbench/common/contributions.js';
-import { Menus } from '../../../../browser/menus.js';
-import { ActiveSessionProviderIdContext, IsPhoneLayoutContext } from '../../../../common/contextkeys.js';
 import { IWorkbenchLayoutService } from '../../../../../workbench/services/layout/browser/layoutService.js';
 import { reportNewChatPickerClosed } from '../../../chat/browser/newChatPickerTelemetry.js';
 import { ISessionsProvidersService } from '../../../../services/sessions/browser/sessionsProvidersService.js';
 import { ISessionsManagementService } from '../../../../services/sessions/common/sessionsManagement.js';
 import type { ISessionsProvider } from '../../../../services/sessions/common/sessionsProvider.js';
-import { type IAgentHostSessionsProvider, isAgentHostProvider, LOCAL_AGENT_HOST_PROVIDER_ID, REMOTE_AGENT_HOST_PROVIDER_RE } from '../../../../common/agentHostSessionsProvider.js';
-import { PermissionPicker } from '../../copilotChatSessions/browser/permissionPicker.js';
-import { MobilePermissionPicker } from '../../copilotChatSessions/browser/mobilePermissionPicker.js';
-import { isPhoneLayout } from '../../../../browser/parts/mobile/mobileLayout.js';
-import { showMobilePickerSheet, IMobilePickerSheetItem, IMobilePickerSheetSearchSource } from '../../../../browser/parts/mobile/mobilePickerSheet.js';
-import { AgentHostModePicker } from './agentHostModePicker.js';
-import { MobileAgentHostModePicker } from './mobile/mobileAgentHostModePicker.js';
-import { AgentHostPermissionPickerDelegate, isWellKnownAutoApproveSchema, isWellKnownModeSchema } from './agentHostPermissionPickerDelegate.js';
+import { type IAgentHostSessionsProvider, isAgentHostProvider } from '../../../../common/agentHostSessionsProvider.js';
+import { isWellKnownAutoApproveSchema, isWellKnownModeSchema } from './agentHostPermissionPickerDelegate.js';
 import { SessionConfigKey } from '../../../../../platform/agentHost/common/sessionConfigKeys.js';
-
-const IsActiveSessionRemoteAgentHost = ContextKeyExpr.regex(ActiveSessionProviderIdContext.key, REMOTE_AGENT_HOST_PROVIDER_RE);
-const IsActiveSessionLocalAgentHost = ContextKeyExpr.equals(ActiveSessionProviderIdContext.key, LOCAL_AGENT_HOST_PROVIDER_ID);
-
-registerAction2(class extends Action2 {
-	constructor() {
-		super({
-			id: 'sessions.agentHost.sessionConfigPicker',
-			title: localize2('agentHostSessionConfigPicker', "Session Configuration"),
-			f1: false,
-			menu: [{
-				id: Menus.NewSessionRepositoryConfig,
-				group: 'navigation',
-				order: 3,
-				when: ContextKeyExpr.or(IsActiveSessionLocalAgentHost, IsActiveSessionRemoteAgentHost),
-			}],
-		});
-	}
-
-	override async run(): Promise<void> { }
-});
 
 export interface IConfigPickerItem {
 	readonly value: string;
@@ -506,358 +471,3 @@ export class AgentHostSessionConfigPicker extends Disposable {
 	}
 }
 
-/**
- * Phone variant of {@link AgentHostSessionConfigPicker} that routes the
- * Isolation and Branch pickers through a unified bottom sheet rather
- * than the desktop action-widget popup.
- *
- * On desktop viewports the inherited `_showPicker` falls through to the
- * base implementation, so this class is safe to keep through
- * viewport-class transitions.
- *
- * Defined in the same file as the base class to avoid a circular ESM
- * dependency (the `extends` clause runs at class-definition time, which
- * is during module evaluation — a separate file that imported the base
- * would hit "Cannot access before initialization").
- */
-class MobileAgentHostSessionConfigPicker extends AgentHostSessionConfigPicker {
-
-	/**
-	 * On phone the chip lane has a fixed visual sequence — Default
-	 * Approvals (rendered by a separate left-side picker), then Branch,
-	 * then Worktree. Sort the known repo-config properties to that
-	 * order; unknown properties fall through to schema-declared order
-	 * after the known ones.
-	 *
-	 * On desktop viewports this subclass is also instantiated (see the
-	 * factory in `AgentHostSessionConfigPickersContribution` — it always
-	 * picks the mobile-aware subclass so `_showPicker` can route to the
-	 * bottom sheet on phones), so we must defer to the base ordering
-	 * (Isolation first, Branch second) when not on a phone layout.
-	 */
-	protected override _orderProperties(properties: ReadonlyArray<[string, SessionConfigPropertySchema]>): ReadonlyArray<[string, SessionConfigPropertySchema]> {
-		if (!isPhoneLayout(this._layoutService)) {
-			return super._orderProperties(properties);
-		}
-		const order = new Map<string, number>([
-			[SessionConfigKey.Branch, 0],
-			[SessionConfigKey.Isolation, 1],
-		]);
-		return properties.slice().sort(([aKey], [bKey]) => {
-			const a = order.get(aKey) ?? Number.MAX_SAFE_INTEGER;
-			const b = order.get(bKey) ?? Number.MAX_SAFE_INTEGER;
-			return a - b;
-		});
-	}
-
-	/**
-	 * Keep Branch and Isolation visible in running sessions even when
-	 * the schema marks them non-mutable. Their value is informational
-	 * — the user wants to see what the running session is using —
-	 * and the chip renders as readonly via {@link _isReadOnlyChip}.
-	 * All other properties defer to the base behavior (hide if
-	 * non-mutable in a running session).
-	 */
-	protected override _shouldRenderProperty(property: string, schema: SessionConfigPropertySchema, isNewSession: boolean): boolean {
-		const isUnifiedRepoProperty = property === SessionConfigKey.Isolation || property === SessionConfigKey.Branch;
-		return isUnifiedRepoProperty || super._shouldRenderProperty(property, schema, isNewSession);
-	}
-
-	/**
-	 * Mark non-mutable properties as readonly chips in running sessions
-	 * so taps don't try to open a picker (which would no-op at the
-	 * provider boundary). The schema's own `readOnly` flag still wins.
-	 */
-	protected override _isReadOnlyChip(property: string, schema: SessionConfigPropertySchema, isNewSession: boolean): boolean {
-		return super._isReadOnlyChip(property, schema, isNewSession) || (!isNewSession && !schema.sessionMutable);
-	}
-
-	protected override async _showPicker(provider: IAgentHostSessionsProvider, sessionId: string, property: string, schema: SessionConfigPropertySchema, trigger: HTMLElement): Promise<void> {
-		if (!isPhoneLayout(this._layoutService)) {
-			return super._showPicker(provider, sessionId, property, schema, trigger);
-		}
-
-		if (property === SessionConfigKey.Isolation || property === SessionConfigKey.Branch) {
-			await this._showUnifiedRepoSheet(provider, sessionId, trigger);
-			return;
-		}
-
-		return super._showPicker(provider, sessionId, property, schema, trigger);
-	}
-
-	private async _showUnifiedRepoSheet(provider: IAgentHostSessionsProvider, sessionId: string, trigger: HTMLElement): Promise<void> {
-		const config = provider.getSessionConfig(sessionId);
-		if (!config) {
-			return;
-		}
-
-		const isolationSchema = config.schema.properties[SessionConfigKey.Isolation];
-		const branchSchema = config.schema.properties[SessionConfigKey.Branch];
-
-		const [isolationItems, branchItems] = await Promise.all([
-			isolationSchema && !isolationSchema.readOnly
-				? this._getItems(provider, sessionId, SessionConfigKey.Isolation, isolationSchema)
-				: Promise.resolve([] as readonly IConfigPickerItem[]),
-			branchSchema && !branchSchema.readOnly
-				? this._getItems(provider, sessionId, SessionConfigKey.Branch, branchSchema)
-				: Promise.resolve([] as readonly IConfigPickerItem[]),
-		]);
-
-		const isolationValue = config.values[SessionConfigKey.Isolation];
-		const branchValue = config.values[SessionConfigKey.Branch];
-		const sheetItems: IMobilePickerSheetItem[] = [];
-
-		const idToConfig = new Map<string, { property: string; value: string; label: string; isPII: boolean }>();
-		const registerId = (property: string, value: string, label: string, isPII: boolean): string => {
-			const id = `repo-row-${idToConfig.size}`;
-			idToConfig.set(id, { property, value, label, isPII });
-			return id;
-		};
-
-		isolationItems.forEach((item, index) => {
-			sheetItems.push({
-				id: registerId(SessionConfigKey.Isolation, item.value, item.label, !!isolationSchema?.enumDynamic),
-				label: item.label,
-				description: item.description,
-				icon: getConfigIcon(SessionConfigKey.Isolation, item.value),
-				checked: item.value === isolationValue,
-				sectionTitle: index === 0 ? (isolationSchema?.title ?? localize('mobileAgentHostSessionConfig.repoSheet.isolationSection', "Isolation")) : undefined,
-			});
-		});
-
-		const branchSectionTitle = branchSchema?.title ?? localize('mobileAgentHostSessionConfig.repoSheet.branchSection', "Base Branch");
-		if (!branchSchema?.enumDynamic) {
-			branchItems.forEach((item, index) => {
-				sheetItems.push({
-					id: registerId(SessionConfigKey.Branch, item.value, item.label, !!branchSchema?.enumDynamic),
-					label: item.label,
-					description: item.description,
-					icon: getConfigIcon(SessionConfigKey.Branch, item.value),
-					checked: item.value === branchValue,
-					sectionTitle: index === 0 ? branchSectionTitle : undefined,
-				});
-			});
-		}
-
-		if (sheetItems.length === 0 && !branchSchema?.enumDynamic) {
-			return;
-		}
-
-		let search: IMobilePickerSheetSearchSource | undefined;
-		if (branchSchema?.enumDynamic && !branchSchema.readOnly) {
-			search = {
-				placeholder: localize('mobileAgentHostSessionConfig.repoSheet.branchSearchPlaceholder', "Search branches"),
-				ariaLabel: localize('mobileAgentHostSessionConfig.repoSheet.branchSearchAria', "Search base branches"),
-				resultsSectionTitle: branchSectionTitle,
-				emptyMessage: localize('mobileAgentHostSessionConfig.repoSheet.branchSearchEmpty', "No matching branches."),
-				loadItems: async (query, token) => {
-					const items = query
-						? await this._getItems(provider, sessionId, SessionConfigKey.Branch, branchSchema, query)
-						: branchItems;
-					if (token.isCancellationRequested) {
-						return [];
-					}
-					return items.map(item => ({
-						id: registerId(SessionConfigKey.Branch, item.value, item.label, !!branchSchema.enumDynamic),
-						label: item.label,
-						description: item.description,
-						icon: getConfigIcon(SessionConfigKey.Branch, item.value),
-						checked: item.value === branchValue,
-					}));
-				},
-			};
-		}
-
-		trigger.setAttribute('aria-expanded', 'true');
-		await showMobilePickerSheet(
-			this._layoutService.mainContainer,
-			localize('mobileAgentHostSessionConfig.repoSheet.title', "Worktree"),
-			sheetItems,
-			{
-				search,
-				// Keep the sheet open on row taps so the user can adjust
-				// both isolation mode and branch without reopening. Each
-				// tap writes through immediately; Done just dismisses.
-				stayOpenOnSelect: true,
-				onDidSelect: (id) => {
-					const selection = idToConfig.get(id);
-					if (selection) {
-						const beforeValue = provider.getSessionConfig(sessionId)?.values[selection.property];
-						reportNewChatPickerClosed(this._telemetryService, {
-							id: 'NewChatAgentHostSessionConfigPicker',
-							name: `NewChatAgentHostSessionConfigPicker.${selection.property}`,
-							optionIdBefore: typeof beforeValue === 'string' ? beforeValue : undefined,
-							optionIdAfter: selection.value,
-							optionLabelBefore: undefined,
-							optionLabelAfter: selection.label,
-							isPII: selection.isPII,
-						});
-						provider.setSessionConfigValue(sessionId, selection.property, selection.value).catch(() => { /* best-effort */ });
-					}
-				},
-			},
-		);
-		trigger.setAttribute('aria-expanded', 'false');
-		trigger.focus();
-	}
-}
-
-interface IConfigPickerWidget extends IDisposable {
-	render(container: HTMLElement): void;
-}
-
-export class PickerActionViewItem extends BaseActionViewItem {
-	constructor(private readonly _picker: IConfigPickerWidget, disposable?: IDisposable) {
-		super(undefined, { id: '', label: '', enabled: true, class: undefined, tooltip: '', run: () => { } });
-		if (disposable) {
-			this._register(disposable);
-		}
-	}
-
-	override render(container: HTMLElement): void {
-		this._picker.render(container);
-	}
-
-	override dispose(): void {
-		this._picker.dispose();
-		super.dispose();
-	}
-}
-
-class AgentHostSessionConfigPickerContribution extends Disposable implements IWorkbenchContribution {
-	static readonly ID = 'sessions.contrib.agentHostSessionConfigPicker';
-
-	constructor(
-		@IActionViewItemService actionViewItemService: IActionViewItemService,
-		@IInstantiationService private readonly _instantiationService: IInstantiationService,
-		@IWorkbenchLayoutService private readonly _layoutService: IWorkbenchLayoutService,
-	) {
-		super();
-		// The mode-picker factories below pick the mobile subclass at
-		// view-item construction time when the viewport is phone, and
-		// the desktop class otherwise. The session-config picker
-		// always uses the mobile-aware subclass because its
-		// `_showPicker` override falls back to `super._showPicker()`
-		// on desktop. The static import of `MobileAgentHostModePicker`
-		// / `MobileAgentHostSessionConfigPicker` creates a circular
-		// dependency (mobile → base → mobile), but ESM handles it
-		// because the classes are only accessed inside these factory
-		// callbacks, which run at `AfterRestored` — well after both
-		// modules have finished evaluating.
-		this._register(actionViewItemService.register(
-			Menus.NewSessionRepositoryConfig,
-			'sessions.agentHost.sessionConfigPicker',
-			() => new PickerActionViewItem(this._instantiationService.createInstance(MobileAgentHostSessionConfigPicker)),
-		));
-		this._register(actionViewItemService.register(
-			Menus.NewSessionConfig,
-			NEW_SESSION_MODE_PICKER_ID,
-			() => new PickerActionViewItem(this._instantiationService.createInstance(
-				isPhoneLayout(this._layoutService) ? MobileAgentHostModePicker : AgentHostModePicker,
-			)),
-		));
-		this._register(actionViewItemService.register(
-			MenuId.ChatInput,
-			RUNNING_SESSION_MODE_PICKER_ID,
-			() => new PickerActionViewItem(this._instantiationService.createInstance(
-				isPhoneLayout(this._layoutService) ? MobileAgentHostModePicker : AgentHostModePicker,
-			)),
-		));
-		this._register(actionViewItemService.register(
-			Menus.NewSessionControl,
-			NEW_SESSION_APPROVE_PICKER_ID,
-			() => this._createNewSessionPermissionPicker(),
-		));
-	}
-
-	/**
-	 * On the new-chat page (left of the toolbar), use the sessions
-	 * {@link PermissionPicker} so the styling matches the surrounding sessions
-	 * pickers (font size, padding, icon size).
-	 */
-	private _createNewSessionPermissionPicker(): PickerActionViewItem {
-		const delegate = this._instantiationService.createInstance(AgentHostPermissionPickerDelegate);
-		const picker = this._instantiationService.createInstance(MobilePermissionPicker, delegate);
-		return new PickerActionViewItem(picker, delegate);
-	}
-}
-
-// ---- New session auto-approve picker (left side, NewSessionControl) ----
-
-const NEW_SESSION_APPROVE_PICKER_ID = 'sessions.agentHost.newSessionApprovePicker';
-
-registerAction2(class extends Action2 {
-	constructor() {
-		super({
-			id: NEW_SESSION_APPROVE_PICKER_ID,
-			title: localize2('agentHostNewSessionApprovePicker', "Session Approvals"),
-			f1: false,
-			menu: [{
-				id: Menus.NewSessionControl,
-				group: 'navigation',
-				order: 1,
-				when: ContextKeyExpr.or(IsActiveSessionLocalAgentHost, IsActiveSessionRemoteAgentHost),
-			}],
-		});
-	}
-
-	override async run(): Promise<void> { }
-});
-
-
-// ---- New session mode picker (NewSessionConfig) ----
-
-const NEW_SESSION_MODE_PICKER_ID = 'sessions.agentHost.newSessionModePicker';
-
-registerAction2(class extends Action2 {
-	constructor() {
-		super({
-			id: NEW_SESSION_MODE_PICKER_ID,
-			title: localize2('agentHostNewSessionModePicker', "Agent Mode"),
-			f1: false,
-			menu: [{
-				id: Menus.NewSessionConfig,
-				group: 'navigation',
-				order: 0,
-				// On phone the {@link MobileChatInputConfigPicker} replaces
-				// this picker with a unified mode + model bottom sheet, so
-				// gate this desktop-only Action out of phone layouts.
-				when: ContextKeyExpr.and(
-					ContextKeyExpr.or(IsActiveSessionLocalAgentHost, IsActiveSessionRemoteAgentHost),
-					IsPhoneLayoutContext.negate(),
-				),
-			}],
-		});
-	}
-
-	override async run(): Promise<void> { }
-});
-
-
-// ---- Running session mode picker (ChatInput, beside the model picker) ----
-
-const RUNNING_SESSION_MODE_PICKER_ID = 'sessions.agentHost.runningSessionModePicker';
-
-registerAction2(class extends Action2 {
-	constructor() {
-		super({
-			id: RUNNING_SESSION_MODE_PICKER_ID,
-			title: localize2('agentHostRunningSessionModePicker', "Agent Mode"),
-			f1: false,
-			menu: [{
-				id: MenuId.ChatInput,
-				group: 'navigation',
-				// `OpenModelPickerAction` (the "Auto" model picker) is at order 3
-				// in the same menu — sit just before it so the mode pill renders
-				// to the left of "Pick Model".
-				order: 2,
-				when: ChatContextKeyExprs.isAgentHostSession,
-			}],
-		});
-	}
-
-	override async run(): Promise<void> { }
-});
-
-
-registerWorkbenchContribution2(AgentHostSessionConfigPickerContribution.ID, AgentHostSessionConfigPickerContribution, WorkbenchPhase.AfterRestored);
